refactor(qa): extract scenario definitions into tools/regression_scenarios.py
